[TREEAPI] remove any logic from VARPs default values
preprocessor replacements are ok though
